add new files xogame, add game function in there and import styles, delet Home file
