+++ README.md
@@ -1,3 +1,11 @@
+## Minigame Collection
+
+The Site is a collection of minigames for one player and intended as a little pastime.
+
+## Memory
+
+The Memory game is palyed on a 4x4 grid. The player uncover cards and looking for matching pairs.
+
 ## X and O Game
 
 The X and O Game is played on a 3×3 grid. The player who has the first move is X. The player who plays second is O. The first player to occupy three spaces in a row, column, or diagonal wins.
