--- components/Home.js
@@ -1,15 +1,0 @@
-import styled from "styled-components";
-
-const Main = styled.main`
-  padding: 0 2rem;
-  min-height: 100vh;
-  padding: 4rem 0;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-`;
-
-export default Main;
