+++ package.json
@@ -13,6 +13,7 @@
     "next": "12.3.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "router": "^1.3.7",
     "styled-components": "^5.3.5"
   },
   "devDependencies": {
