+++ pages/index.js
@@ -1,23 +1,22 @@
 import Head from "next/head";
 import Header from "../components/Header";
 import XOGame from "../components/XOGame";
-import Main from "../components/Home";
+import Link from "next/link";
+import { useRouter } from "next/router";
 
 
 export default function Home() {
+  const router = useRouter();
+  const path = router.pathname;
   return (
     <>
       <Head>
-        <title>X O Game</title>
+        <title>Minigame Collection</title>
         <meta name="description" content="Generated by create next app" />
-        <link rel="shortcut icon" type="image/x-icon" href="./favicon.ico" />       
+        <link rel="shortcut icon" type="image/x-icon" href="./favicon.ico" />
       </Head>
-      
-      <Main>
-        <Header />
-        <XOGame />
-        
-      </Main>
+      <Header />
+      <Link href="/xogame">X and O Game</Link>
     </>
   )
 }
